uploading the scripts i developped in the last week for my virtual reality headset idea.
I will be using those inside of my virtual reality desktop projects but not necessarily all of them. I wanted to upload these as i was proud of achieving  this although i am not sure from other programmers perspectives that creating sockets every frame is a good idea... but it works for me and hopefully it will work for others too. I am not using any threads for the moment as i do not know how to use them yet in python but soon i will.

steve chassé aka ninekorn